refactor of messaging schema and controller
